+++ src/azure-firewall/HISTORY.rst
@@ -3,6 +3,24 @@
 Release History
 ===============
 
+0.2.0
+++++++
+* `az network firewall ip-config`: deprecate --private-ip-address
+* `az network firewall create/update`: support --sku, --firewall-policy and --vhub.
+
+0.1.9
+++++++
+* `az network firewall management-ip-config`: support creating/showing management ip configuration.
+
+0.1.8
+++++++
+* `az network firewall create`: support private-ranges
+* `az network firewall threat-intel-whitelist`:  support threat intelligence whitelist.
+
+0.1.7
+++++++
+* `az network firewall network-rule/nat-rule/application-rule`: support ip-groups for firewall rules.
+
 0.1.6
 ++++++
 * `az network firewall ip-config`: bug fix to support multiple ip-config creation.